Move File related methods into FileUtil
