+++ conversions/Forwards/src/main/java/com/agentdid127/resourcepack/forwards/impl/SoundsConverter.java
@@ -3,9 +3,9 @@ package com.agentdid127.resourcepack.forwards.impl;
 import com.agentdid127.resourcepack.library.Converter;
 import com.agentdid127.resourcepack.library.PackConverter;
 import com.agentdid127.resourcepack.library.pack.Pack;
+import com.agentdid127.resourcepack.library.utilities.FileUtil;
 import com.agentdid127.resourcepack.library.utilities.JsonUtil;
 import com.agentdid127.resourcepack.library.utilities.Logger;
-import com.agentdid127.resourcepack.library.utilities.Util;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -58,7 +58,7 @@ public class SoundsConverter extends Converter {
                         Path baseSoundsPath = pack.getWorkingPath()
                                 .resolve("assets/minecraft/sounds".replace("/", File.separator));
                         Path path = baseSoundsPath.resolve(sound + ".ogg");
-                        if (!Util.fileExistsCorrectCasing(path)) {
+                        if (!FileUtil.fileExistsCorrectCasing(path)) {
                             String rewrite = path.toFile().getCanonicalPath().substring(
                                     baseSoundsPath.toString().length() + 1,
                                     path.toFile().getCanonicalPath().length() - 4);
